split attach_cve_flaw lib
